+++ src/client/CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to
 ### Added
 
 - Add tariff API client endpoint and CLI commands
+- Add `original_id` filtering to tariff client list calls
 
 #### Dependencies